Fixed some compiler warnings.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/900



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@4393 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/test/intltest/normconf.cpp
@@ -86,7 +86,7 @@ void NormalizerConformanceTest::TestConformance(void) {
         if (T_FileStream_eof(input)) {
             break;
         }
-        T_FileStream_readLine(input, lineBuf, sizeof(lineBuf));
+        T_FileStream_readLine(input, lineBuf, (int32_t)sizeof(lineBuf));
         UnicodeString line(lineBuf, "");
         if (line.length() == 0) continue;
